Address Decoder GAL files
I'm using a GAL22V10, programmed with GALBLAST via the parallel port on an old Win98 machine.

The PIT (Programmable Interval Timer) mentioned in the ADDECODE.PLD file was never used, and that output (connected to EXSEL1 on the PCB) is available as an active low select line for the addresses $7F40-7F7F.  The EXSEL2 on the PCB is currently configured as an active high reset.  Both of these can be reconfigured, if desired, as they are not currently used by any of the current hardware.